Tidied up config
